TST: GH26807 move a couple more tests out of test_store
